% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_utilities.R
\name{close_logging_progress_bar}
\alias{close_logging_progress_bar}
\title{Close a logging progress bar}
\usage{
close_logging_progress_bar(pb, ...)
}
\arguments{
\item{pb}{A logging progress bar object}

\item{...}{Additional arguments (not used)}
}
\description{
Close a logging progress bar
}
