% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corandcov}
\alias{corandcov}
\title{Extract working correlation matrix from a \code{gls}/\code{lme} fit}
\usage{
corandcov(glsob, cov = TRUE, ...)
}
\arguments{
\item{glsob}{A \code{gls} or \code{lme} object with a structured correlation
component.}

\item{cov}{Logical; if \code{TRUE}, return covariance (default).}

\item{...}{Reserved for future use.}
}
\value{
A correlation matrix.
}
\description{
Extract working correlation matrix from a \code{gls}/\code{lme} fit
}
