% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_timepoint_names}
\alias{get_timepoint_names}
\title{Return the timepoint-name column of a trial-design tibble}
\usage{
get_timepoint_names(td)
}
\arguments{
\item{td}{Trial-design data frame or tibble.}
}
\value{
Character vector of timepoint names.
}
\description{
Return the timepoint-name column of a trial-design tibble
}
\keyword{internal}
