% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_utilities.R
\name{logging_progress_bar}
\alias{logging_progress_bar}
\title{Set up a text progress bar that can also log to a file}
\usage{
logging_progress_bar(
  min = 0,
  max = 1,
  initial = 0,
  style = 3,
  log_file = NULL,
  log_frequency = 10
)
}
\arguments{
\item{min}{Minimum value for the progress bar}

\item{max}{Maximum value for the progress bar}

\item{initial}{Initial value for the progress bar}

\item{style}{Style of the progress bar (see txtProgressBar)}

\item{log_file}{Optional file path to write progress updates to}

\item{log_frequency}{How often to log to file (in terms of update count)}
}
\value{
A progress bar object with additional logging capabilities
}
\description{
This function creates a progress bar similar to txtProgressBar but with
the option to log progress updates to a file.
}
