% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carryover_sensitivity.R
\name{plot_carryover_models}
\alias{plot_carryover_models}
\title{Create comparison plot of carryover models}
\usage{
plot_carryover_models(
  time_range = seq(0, 10, by = 0.1),
  initial_effect = 10,
  model_params = list(exponential = list(carryover_t1half = 2, scalefactor = 2), linear =
    list(total_time = 8), weibull = list(carryover_t1half = 2, k = 1.5), power =
    list(max_time = 8, power = 1.5)),
  models = c("exponential", "linear", "weibull", "power")
)
}
\arguments{
\item{time_range}{Vector of time values to plot}

\item{initial_effect}{Initial effect value at time 0}

\item{model_params}{List of model parameters for each model type}

\item{models}{Vector of models to include}
}
\value{
invisible(NULL)
}
\description{
Visualizes different carryover models on the same plot to compare
their decay profiles over time.
}
