% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utilities.R
\name{prepare_parallel_globals}
\alias{prepare_parallel_globals}
\title{Create global function list for parallel processing}
\usage{
prepare_parallel_globals(
  functions = NULL,
  packages = NULL,
  add_defaults = TRUE
)
}
\arguments{
\item{functions}{Vector of function names to export as globals}

\item{packages}{Vector of package names to load in workers}

\item{add_defaults}{Whether to include a default set of commonly used functions}
}
\value{
List of functions for use with future::plan globals parameter
}
\description{
This function creates a list of global functions that are needed in
parallel processes. It ensures that all necessary functions are available
to workers in future/furrr parallel processing.
}
