% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{resolve_conflicts}
\alias{resolve_conflicts}
\title{Resolve common conflicts}
\usage{
resolve_conflicts()
}
\value{
NULL, invisibly
}
\description{
This function resolves common package conflicts in the nof1power package.
It explicitly declares preferences for certain functions when multiple packages
provide functions with the same name.
}
