% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utilities.R
\name{setup_vignette_parallel}
\alias{setup_vignette_parallel}
\title{Configure vignette for parallel processing}
\usage{
setup_vignette_parallel(
  load_full_package = TRUE,
  functions = NULL,
  verbose = TRUE
)
}
\arguments{
\item{load_full_package}{Whether to attempt a devtools::load_all() operation}

\item{functions}{Vector of function names that should be made available to workers}

\item{verbose}{Whether to print status messages}
}
\value{
Invisibly returns the future plan object
}
\description{
This function is specifically designed to be called at the top of vignettes
to ensure all needed functions are available to workers when using
future/furrr for parallel processing.
}
