% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_utilities.R
\name{update_logging_progress_bar}
\alias{update_logging_progress_bar}
\title{Update a logging progress bar}
\usage{
update_logging_progress_bar(pb, value, ...)
}
\arguments{
\item{pb}{A logging progress bar object}

\item{value}{Current value}

\item{...}{Additional arguments (not used)}
}
\value{
The progress bar object, invisibly
}
\description{
Update a logging progress bar
}
